+++ packages/helldoc-cli/src/index.ts
@@ -1,22 +1,26 @@
 #!/usr/bin/env node
 
 import * as program from "commander";
-import { dev, prod } from "@helldoc/core";
+import { prod, dev } from "@helldoc/core";
 
 program.version("0.0.1");
-program
-  .command("dev <dir>")
+
+bindUniversalOptions(program.command("dev <dir>"))
   .description("run the docs dev server with dir path")
-  .action(function(dir, _options) {
-    dev(dir);
+  .action(function(dir, options) {
+    dev(dir, options);
   });
-program
-  .command("prod <dir>")
+
+bindUniversalOptions(program.command("prod <dir>"))
   .description("output the static web files with dir path")
-  .action(function(dir, _options) {
-    prod(dir);
+  .action(function(dir, options) {
+    prod(dir, options);
   });
 
 program.parse(process.argv);
 
 if (!program.args.length) program.help();
+
+function bindUniversalOptions(program: program.Command) {
+  return program.option("-o, --output <dir>", "set output dir path.");
+}