+++ packages/helldoc-core/src/prod.ts
@@ -1,23 +1,24 @@
 import * as Webpack from "webpack";
 
-import prepare from "./prepare";
+import prepare, { CLIOptions } from "./prepare";
 import createClientConfig from "./webpack/createClientConfig";
 
-async function prod(sourceDir: string) {
+async function prod(sourceDir: string, cliOptions: CLIOptions) {
   process.env.NODE_ENV = "production";
 
   console.log("\nExtracting site metadata...");
-  const options = await prepare(sourceDir);
+  const options = await prepare(sourceDir, cliOptions);
 
   const configChain = createClientConfig(options);
   const config = configChain.toConfig();
   const compiler = Webpack(config);
 
-  compiler.run(err => {
+  compiler.run((err, stats) => {
+    console.log(err, stats.compilation.errors);
     if (err) {
       console.error(err);
     }
-    console.log("\nProd site done. serve the /dist make a try.");
+    console.log("\nProd site done.");
   });
 }